Timezone naming change in course detail page

+++ src/course-home/outline-tab/SequenceLink.jsx
@@ -51,7 +51,7 @@ const SequenceLink = ({
             day="numeric"
             month="short"
             year="numeric"
-            timeZoneName="short"
+            timeZoneName="long"
             value={due}
             {...timezoneFormatArgs}
           />
@@ -73,7 +73,7 @@ const SequenceLink = ({
             day="numeric"
             month="short"
             year="numeric"
-            timeZoneName="short"
+            timeZoneName="long"
             value={due}
             {...timezoneFormatArgs}
           />
